+++ src/components/common/index.js
@@ -6,11 +6,13 @@ import Text from './text'
 import FramedIcon from './framed-icon'
 import Widget from './widget'
 import Toast from './toast'
+import Link from './link'
 
 export {
   Button,
   Toast,
   Loader,
+  Link,
   Title,
   TokenImage,
   Text,