+++ src/components/pages/common/wallets-list/index.tsx
@@ -17,6 +17,7 @@ import ZerionWalletIcon from 'images/zerion-wallet.png'
 import RainbowWalletIcon from 'images/rainbow-wallet.png'
 import ImtokenWalletIcon from 'images/imtoken-wallet.png'
 import WalletConnectIcon from 'images/walletconnect-wallet.png'
+import CrossmintIcon from 'images/crossmint-wallet.png'
 import ENSIcon from 'images/ens-logo.png'
 import { useConnect, Connector } from 'wagmi'
 import { TDropStep, TMultiscanStep, TWalletName, TWalletOption } from 'types'
@@ -95,7 +96,7 @@ const defineOptionsList = (
   open: (options?: any | undefined) => Promise<void>,
   connect: (args: Partial<any> | undefined) => void,
   connectors: Connector<any, any, any>[],
-  wallet: string | null,
+  wallet: TWalletName | null,
   deeplinkRedirect: (
     deeplink: string,
     walletId: TWalletName
@@ -108,7 +109,7 @@ const defineOptionsList = (
 
   const system = defineSystem()
   const ensOption = !isManual && enableENS ? {
-    title: 'Enter ENS or address',
+    title: 'ENS or address',
     onClick: () => setStep('set_address'),
     icon: <WalletIcon src={ENSIcon} />
   } : undefined
@@ -121,6 +122,16 @@ const defineOptionsList = (
     icon: <WalletIcon src={WalletConnectIcon} />,
     recommended: wallet === 'walletconnect'
   }
+
+  const crossmintOption = {
+    title: 'Crossmint',
+    onClick: () => {
+      setStep('crossmint_connection')
+    },
+    icon: <WalletIcon src={CrossmintIcon} />,
+    recommended: wallet === 'crossmint'
+  }
+
   const injected = connectors.find(connector => connector.id === "injected")
   const injectedOption = getInjectedWalletOption(
     wallet,
@@ -147,9 +158,10 @@ const defineOptionsList = (
 
     const wallets = [
       isOptionVisible(injectedOption, wallet, 'metamask', availableWallets),
+      isOptionVisible(crossmintOption, wallet, 'crossmint', availableWallets),
       isOptionVisible(coinbaseOption, wallet, 'coinbase_wallet', availableWallets),
       isOptionVisible(walletConnectOption, wallet, 'walletconnect', availableWallets),
-      ensOption
+      isOptionVisible(ensOption, wallet, 'manual_address', availableWallets)
     ]
 
     return sortWallets(wallets) 
@@ -228,7 +240,8 @@ const defineOptionsList = (
     isOptionVisible(coinbaseOption, wallet, 'coinbase_wallet', availableWallets),
     isOptionVisible(zerionOption, wallet, 'zerion', availableWallets),
     isOptionVisible(walletConnectOption, wallet, 'walletconnect', availableWallets),
-    ensOption,
+    isOptionVisible(crossmintOption, wallet, 'crossmint', availableWallets),
+    isOptionVisible(ensOption, wallet, 'manual_address', availableWallets),
     isOptionVisible(imtokenOption, wallet, 'imtoken', availableWallets),
     isOptionVisible(trustOption, wallet, 'trust', availableWallets),
     isOptionVisible(rainbowOption, wallet, 'rainbow', availableWallets)
@@ -265,6 +278,7 @@ const WalletsList: FC<ReduxType> = ({
     availableWallets,
     enableENS
   )
+
 
   return <Container>
     <TitleComponent>Connect your wallet</TitleComponent>