+++ src/data/store/reducers/user/async-actions/index.js
@@ -1,9 +1,11 @@
 import initialize from "./initialize"
 import switchNetwork from "./switch-network"
 import updateUserData from "./update-user-data"
+import getCrossmintAddress from "./get-crossmint-address"
 
 export {
   initialize,
   switchNetwork,
-  updateUserData
+  updateUserData,
+  getCrossmintAddress
 }